django-backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
black==26.1.0
cbor2==5.8.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.13.4
cross-web==0.4.1
cryptography==46.0.5
daphne==4.2.1
Django==5.2.10
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular
factory_boy==3.3.3
Faker==40.4.0
graphql-core==3.2.7
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
mnemonic==0.21
msgpack==1.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyNaCl==1.6.2
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-asyncio==0.25.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytokens==0.4.0
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
requests-sse==0.5.2
responses==0.26.0
rpds-py==0.30.0
ruff==0.14.13
sentry-sdk==2.53.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
stellar-sdk==13.2.0
strawberry-graphql[channels]==0.289.0
strawberry-graphql-django==0.74.1
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
xdrlib3==0.1.1
zope.interface==8.2